Add files via upload
MLY90393 is a magnetometer. In order to use it, it must be first calibrated. This is done by running the calibration file generation program, which creates a text file with X Y Z data in nT. (nanoTeslas) The magneto program reads this text file and calculates the correction factor. These correction factors then are corrected into an arduino program, which outputs the corrected X Y Z data. This program runs in the Arduino, which is connected with USB cable to the computer and the program writes the signals from the magnetometer to the COM port. In the computer, the python program reads the data coming from the COM port . Normally I would start the program from VS Code, and out-comment different code lines, to test X Y  Z seperately. and for this seperate testing, the python program only shows one direction at a time as a clock visor on the graphic screen opening. This is, because I got confused, if I run the 3D type of output. The reason for my confusion was, that  te calibration showed, that in my specific Magnetometer, the X, Y signals were smaller (in the range on 15%)  that in the X Z and Y Z direction.